producer/requirements.txt
hazelcast-python-client==5.2.0
kafka-python==2.0.2
